// Portal.Portal.Web/src/app/shared/infrastructure/PortalHttpClient/model/employeeUniversityModel.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { EmployeeFileModel } from './employeeFileModel';

export interface EmployeeUniversityModel { 
    id?: number;
    university?: string;
    faculty?: string;
    startDate?: Date;
    endDate?: Date;
    cerificate?: EmployeeFileModel;
}